Final updates and cleanup

+++ init.sql
@@ -3,7 +3,7 @@ CREATE TABLE IF NOT EXISTS products (
   product_name TEXT NOT NULL,
   category TEXT NOT NULL,
   rating FLOAT DEFAULT 0,
-  review_count INT DEFAULT 0,
+  review_count FLOAT DEFAULT 0,
   discount FLOAT DEFAULT 0,
   price FLOAT DEFAULT 0,
   created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP
